refactor(components.product): move to webUniverseComponent
