+++ client/app/app.js
@@ -1,4 +1,6 @@
 import translateAsyncLoader from '@ovh-ux/translate-async-loader';
+import uiRouter from '@uirouter/angularjs';
+
 import config from './config/config';
 import webUniverseComponents from './webUniverseComponents';
 // TODO: web-universe-components is not yet published;
@@ -32,7 +34,7 @@ angular
     'xeditable',
     'ovh-angular-sso-auth',
     'ovh-angular-sso-auth-modal-plugin',
-    'ui.router',
+    uiRouter,
     'ovh-angular-sidebar-menu',
     'pascalprecht.translate',
     'ovh-angular-responsive-tabs',
