+++ package.json
@@ -35,7 +35,6 @@
     "@bower_components/URIjs": "medialize/URI.js#^1.14.0",
     "@bower_components/angular-route": "angular/bower-angular-route#^1.6.0",
     "@bower_components/angular-scroll-glue": "Luegg/angularjs-scroll-glue#^2.2.0",
-    "@bower_components/angular-ui-router": "angular-ui/angular-ui-router-bower#^1.0.0",
     "@bower_components/angular-vs-repeat": "kamilkp/angular-vs-repeat#v1.1.7",
     "@bower_components/bs4": "twbs/bootstrap#v4.0.0-beta",
     "@bower_components/ckeditor": "ckeditor/ckeditor-releases#^4.7.2",
@@ -48,6 +47,7 @@
     "@bower_components/randexp": "fent/randexp.js#^0.4.0",
     "@ovh-ux/ovh-utils-angular": "^13.0.0",
     "@ovh-ux/translate-async-loader": "^1.0.7",
+    "@uirouter/angularjs": "^1.0.0",
     "angular": "~1.6.10",
     "angular-aria": "~1.6.0",
     "angular-cookies": "~1.6.0",
